chore: powerlaws shifted with truncated distributions
