+++ numpyro/distributions/__init__.py
@@ -20,7 +20,6 @@ from numpyro.distributions.continuous import (
     Cauchy,
     Chi2,
     Dirichlet,
-    DoublyTruncatedPowerLaw,
     EulerMaruyama,
     Exponential,
     Gamma,
@@ -36,7 +35,6 @@ from numpyro.distributions.continuous import (
     Logistic,
     LogNormal,
     LogUniform,
-    LowerTruncatedPowerLaw,
     LowRankMultivariateNormal,
     MatrixNormal,
     MultivariateNormal,
@@ -97,7 +95,9 @@ from numpyro.distributions.kl import kl_divergence
 from numpyro.distributions.mixtures import Mixture, MixtureGeneral, MixtureSameFamily
 from numpyro.distributions.transforms import biject_to
 from numpyro.distributions.truncated import (
+    DoublyTruncatedPowerLaw,
     LeftTruncatedDistribution,
+    LowerTruncatedPowerLaw,
     RightTruncatedDistribution,
     TruncatedCauchy,
     TruncatedDistribution,
@@ -124,6 +124,7 @@ __all__ = [
     "Binomial",
     "BinomialLogits",
     "BinomialProbs",
+    "CAR",
     "Categorical",
     "CategoricalLogits",
     "CategoricalProbs",
@@ -131,13 +132,13 @@ __all__ = [
     "Chi2",
     "Delta",
     "Dirichlet",
-    "DoublyTruncatedPowerLaw",
     "DirichletMultinomial",
     "DiscreteUniform",
     "Distribution",
+    "DoublyTruncatedPowerLaw",
     "EulerMaruyama",
-    "Exponential",
     "ExpandedDistribution",
+    "Exponential",
     "FoldedDistribution",
     "Gamma",
     "GammaPoisson",
@@ -155,30 +156,29 @@ __all__ = [
     "Independent",
     "InverseGamma",
     "Kumaraswamy",
-    "LKJ",
-    "LKJCholesky",
     "Laplace",
     "LeftTruncatedDistribution",
+    "LKJ",
+    "LKJCholesky",
     "Logistic",
     "LogNormal",
     "LogUniform",
     "LowerTruncatedPowerLaw",
-    "MatrixNormal",
+    "LowRankMultivariateNormal",
     "MaskedDistribution",
+    "MatrixNormal",
     "Mixture",
-    "MixtureSameFamily",
     "MixtureGeneral",
+    "MixtureSameFamily",
     "Multinomial",
     "MultinomialLogits",
     "MultinomialProbs",
     "MultivariateNormal",
-    "CAR",
     "MultivariateStudentT",
-    "LowRankMultivariateNormal",
-    "Normal",
-    "NegativeBinomialProbs",
-    "NegativeBinomialLogits",
     "NegativeBinomial2",
+    "NegativeBinomialLogits",
+    "NegativeBinomialProbs",
+    "Normal",
     "OrderedLogistic",
     "Pareto",
     "Poisson",
@@ -203,7 +203,7 @@ __all__ = [
     "Wishart",
     "WishartCholesky",
     "ZeroInflatedDistribution",
-    "ZeroInflatedPoisson",
     "ZeroInflatedNegativeBinomial2",
+    "ZeroInflatedPoisson",
     "ZeroSumNormal",
 ]
