+++ numpyro/distributions/continuous.py
@@ -27,7 +27,6 @@
 
 import numpy as np
 
-import jax
 from jax import lax, vmap
 from jax.experimental.sparse import BCOO
 from jax.lax import scan
@@ -2863,536 +2862,3 @@ class WishartCholesky(Distribution):
                 batch_shape = lax.broadcast_shapes(concentration, matrix[:-2])
                 event_shape = matrix[-2:]
                 return batch_shape, event_shape
-
-
-class DoublyTruncatedPowerLaw(Distribution):
-    r"""Power law distribution with :math:`\alpha` index, and lower and upper bounds.
-    We can define the power law distribution as,
-
-    .. math::
-        f(x; \alpha, a, b) = \frac{x^{\alpha}}{Z(\alpha, a, b)},
-
-    where, :math:`a` and :math:`b` are the lower and upper bounds respectively,
-    and :math:`Z(\alpha, a, b)` is the normalization constant. It is defined as,
-
-    .. math::
-        Z(\alpha, a, b) = \begin{cases}
-            \log(b) - \log(a) & \text{if } \alpha = -1, \\
-            \frac{b^{1 + \alpha} - a^{1 + \alpha}}{1 + \alpha} & \text{otherwise}.
-        \end{cases}
-
-    :param alpha: index of the power law distribution
-    :param low: lower bound of the distribution
-    :param high: upper bound of the distribution
-    """
-
-    arg_constraints = {
-        "alpha": constraints.real,
-        "low": constraints.greater_than_eq(0),
-        "high": constraints.greater_than(0),
-    }
-    reparametrized_params = ["alpha", "low", "high"]
-    pytree_aux_fields = ("_support",)
-    pytree_data_fields = ("alpha", "low", "high")
-
-    def __init__(self, alpha, low, high, *, validate_args=None):
-        self.alpha, self.low, self.high = promote_shapes(alpha, low, high)
-        self._support = constraints.interval(low, high)
-        batch_shape = lax.broadcast_shapes(
-            jnp.shape(alpha), jnp.shape(low), jnp.shape(high)
-        )
-        super(DoublyTruncatedPowerLaw, self).__init__(
-            batch_shape=batch_shape, validate_args=validate_args
-        )
-
-    @constraints.dependent_property(is_discrete=False, event_dim=0)
-    def support(self):
-        return self._support
-
-    @validate_sample
-    def log_prob(self, value):
-        r"""Logarithmic probability distribution:
-        Z inequal minus one:
-        .. math::
-            (x^\alpha) (\alpha + 1)/(b^(\alpha + 1) - a^(\alpha + 1))
-
-        Z equal minus one:
-        .. math::
-            (x^\alpha)/(log(b) - log(a))
-        Derivations are calculated by Wolfram Alpha via the Jacobian matrix accordingly.
-        """
-
-        @jax.custom_jvp
-        def f(x, alpha, low, high):
-            neq_neg1_mask = jnp.not_equal(alpha, -1.0)
-            neq_neg1_alpha = jnp.where(neq_neg1_mask, alpha, 0.0)
-            # eq_neg1_alpha = jnp.where(~neq_neg1_mask, alpha, -1.0)
-
-            def neq_neg1_fn():
-                one_more_alpha = 1.0 + neq_neg1_alpha
-                return jnp.log(
-                    jnp.power(x, neq_neg1_alpha)
-                    * (one_more_alpha)
-                    / (jnp.power(high, one_more_alpha) - jnp.power(low, one_more_alpha))
-                )
-
-            def eq_neg1_fn():
-                return -jnp.log(x) - jnp.log(jnp.log(high) - jnp.log(low))
-
-            return jnp.where(neq_neg1_mask, neq_neg1_fn(), eq_neg1_fn())
-
-        @f.defjvp
-        def f_jvp(primals, tangents):
-            x, alpha, low, high = primals
-            x_t, alpha_t, low_t, high_t = tangents
-
-            log_low = jnp.log(low)
-            log_high = jnp.log(high)
-            log_x = jnp.log(x)
-
-            # Mask and alpha values
-            delta_eq_neg1 = 10e-4
-            neq_neg1_mask = jnp.not_equal(alpha, -1.0)
-            neq_neg1_alpha = jnp.where(neq_neg1_mask, alpha, 0.0)
-            eq_neg1_alpha = jnp.where(jnp.not_equal(alpha, 0.0), alpha, -1.0)
-
-            primal_out = f(*primals)
-
-            # Alpha tangent with approximation
-            # Variable part for all values alpha unequal -1
-            def alpha_tangent_variable(alpha):
-                one_more_alpha = 1.0 + alpha
-                low_pow_one_more_alpha = jnp.power(low, one_more_alpha)
-                high_pow_one_more_alpha = jnp.power(high, one_more_alpha)
-                return jnp.reciprocal(one_more_alpha) + (
-                    low_pow_one_more_alpha * log_low
-                    - high_pow_one_more_alpha * log_high
-                ) / (high_pow_one_more_alpha - low_pow_one_more_alpha)
-
-            # Alpha tangent
-            alpha_tangent = jnp.where(
-                neq_neg1_mask,
-                log_x + alpha_tangent_variable(neq_neg1_alpha),
-                # Approximate derivate with right an lefthand approximation
-                log_x
-                + (
-                    alpha_tangent_variable(alpha - delta_eq_neg1)
-                    + alpha_tangent_variable(alpha + delta_eq_neg1)
-                )
-                * 0.5,
-            )
-
-            # High and low tangents for alpha unequal -1
-            one_more_alpha = 1.0 + neq_neg1_alpha
-            low_pow_one_more_alpha = jnp.power(low, one_more_alpha)
-            high_pow_one_more_alpha = jnp.power(high, one_more_alpha)
-            change_sq = jnp.square(high_pow_one_more_alpha - low_pow_one_more_alpha)
-            low_tangent_neq_neg1_common = (
-                jnp.square(one_more_alpha) * jnp.power(x, neq_neg1_alpha) / change_sq
-            )
-            low_tangent_neq_neg1 = low_tangent_neq_neg1_common * jnp.power(
-                low, neq_neg1_alpha
-            )
-            high_tangent_neq_neg1 = low_tangent_neq_neg1_common * jnp.power(
-                high, neq_neg1_alpha
-            )
-
-            # High and low tangents for alpha equal -1
-            low_tangent_eq_neg1_common = jnp.power(x, eq_neg1_alpha) / jnp.square(
-                log_high - log_low
-            )
-            low_tangent_eq_neg1 = low_tangent_eq_neg1_common / low
-            high_tangent_eq_neg1 = -low_tangent_eq_neg1_common / high
-
-            # High and low tangents
-            low_tangent = jnp.where(
-                neq_neg1_mask, low_tangent_neq_neg1, low_tangent_eq_neg1
-            )
-            high_tangent = jnp.where(
-                neq_neg1_mask, high_tangent_neq_neg1, high_tangent_eq_neg1
-            )
-
-            # Final tangents
-            tangent_out = (
-                alpha / x * x_t
-                + alpha_tangent * alpha_t
-                + low_tangent * low_t
-                + high_tangent * high_t
-            )
-            return primal_out, tangent_out
-
-        return f(value, self.alpha, self.low, self.high)
-
-    def cdf(self, value):
-        r"""Cumulated probability distribution:
-        Z inequal minus one:
-
-        .. math::
-
-            \frac{x^{\alpha + 1} - a^{\alpha + 1}}{b^{\alpha + 1} - a^{\alpha + 1}}
-
-        Z equal minus one:
-
-        .. math::
-
-            \frac{\log(x) - \log(a)}{\log(b) - \log(a)}
-
-        Derivations are calculated by Wolfram Alpha via the Jacobian matrix accordingly.
-        """
-
-        @jax.custom_jvp
-        def f(x, alpha, low, high):
-            neq_neg1_mask = jnp.not_equal(alpha, -1.0)
-            neq_neg1_alpha = jnp.where(neq_neg1_mask, alpha, 0.0)
-
-            def cdf_when_alpha_neq_neg1():
-                one_more_alpha = 1.0 + neq_neg1_alpha
-                low_pow_one_more_alpha = jnp.power(low, one_more_alpha)
-                return (jnp.power(x, one_more_alpha) - low_pow_one_more_alpha) / (
-                    jnp.power(high, one_more_alpha) - low_pow_one_more_alpha
-                )
-
-            def cdf_when_alpha_eq_neg1():
-                return jnp.log(x / low) / jnp.log(high / low)
-
-            cdf_val = jnp.where(
-                neq_neg1_mask,
-                cdf_when_alpha_neq_neg1(),
-                cdf_when_alpha_eq_neg1(),
-            )
-            return jnp.clip(cdf_val, a_min=0.0, a_max=1.0)
-
-        @f.defjvp
-        def f_jvp(primals, tangents):
-            x, alpha, low, high = primals
-            x_t, alpha_t, low_t, high_t = tangents
-
-            log_low = jnp.log(low)
-            log_high = jnp.log(high)
-            log_x = jnp.log(x)
-
-            delta_eq_neg1 = 10e-4
-            neq_neg1_mask = jnp.not_equal(alpha, -1.0)
-            neq_neg1_alpha = jnp.where(neq_neg1_mask, alpha, 0.0)
-
-            # Calculate primal
-            primal_out = f(*primals)
-
-            # Tangents for alpha not equals -1
-            def x_neq_neg1(alpha):
-                one_more_alpha = 1.0 + alpha
-                return (one_more_alpha * jnp.power(x, alpha)) / (
-                    jnp.power(high, one_more_alpha) - jnp.power(low, one_more_alpha)
-                )
-
-            def alpha_neq_neg1(alpha):
-                one_more_alpha = 1.0 + alpha
-                low_pow_one_more_alpha = jnp.power(low, one_more_alpha)
-                high_pow_one_more_alpha = jnp.power(high, one_more_alpha)
-                x_pow_one_more_alpha = jnp.power(x, one_more_alpha)
-                term1 = (
-                    x_pow_one_more_alpha * log_x - low_pow_one_more_alpha * log_low
-                ) / (high_pow_one_more_alpha - low_pow_one_more_alpha)
-                term2 = (
-                    (x_pow_one_more_alpha - low_pow_one_more_alpha)
-                    * (
-                        high_pow_one_more_alpha * log_high
-                        - low_pow_one_more_alpha * log_low
-                    )
-                ) / jnp.square(high_pow_one_more_alpha - low_pow_one_more_alpha)
-                return term1 - term2
-
-            def low_neq_neg1(alpha):
-                one_more_alpha = 1.0 + alpha
-                low_pow_one_more_alpha = jnp.power(low, one_more_alpha)
-                high_pow_one_more_alpha = jnp.power(high, one_more_alpha)
-                x_pow_one_more_alpha = jnp.power(x, one_more_alpha)
-                change = high_pow_one_more_alpha - low_pow_one_more_alpha
-                term2 = one_more_alpha * jnp.power(low, alpha) / change
-                term1 = term2 * (x_pow_one_more_alpha - low_pow_one_more_alpha) / change
-                return term1 - term2
-
-            def high_neq_neg1(alpha):
-                one_more_alpha = 1.0 + alpha
-                low_pow_one_more_alpha = jnp.power(low, one_more_alpha)
-                high_pow_one_more_alpha = jnp.power(high, one_more_alpha)
-                x_pow_one_more_alpha = jnp.power(x, one_more_alpha)
-                return -(
-                    one_more_alpha
-                    * jnp.power(high, alpha)
-                    * (x_pow_one_more_alpha - low_pow_one_more_alpha)
-                ) / jnp.square(high_pow_one_more_alpha - low_pow_one_more_alpha)
-
-            # Tangents for alpha equals -1
-            def x_eq_neg1():
-                return jnp.reciprocal(x * (log_high - log_low))
-
-            def low_eq_neg1():
-                return (log_x - log_low) / (
-                    jnp.square(log_high - log_low) * low
-                ) - jnp.reciprocal((log_high - log_low) * low)
-
-            def high_eq_neg1():
-                return (log_x - log_low) / (jnp.square(log_high - log_low) * high)
-
-            # Inlcuding approximation for alpha = -1
-            tangent_out = (
-                jnp.where(neq_neg1_mask, x_neq_neg1(neq_neg1_alpha), x_eq_neg1()) * x_t
-                + jnp.where(
-                    neq_neg1_mask,
-                    alpha_neq_neg1(neq_neg1_alpha),
-                    (
-                        alpha_neq_neg1(alpha - delta_eq_neg1)
-                        + alpha_neq_neg1(alpha + delta_eq_neg1)
-                    )
-                    * 0.5,
-                )
-                * alpha_t
-                + jnp.where(neq_neg1_mask, low_neq_neg1(neq_neg1_alpha), low_eq_neg1())
-                * low_t
-                + jnp.where(
-                    neq_neg1_mask, high_neq_neg1(neq_neg1_alpha), high_eq_neg1()
-                )
-                * high_t
-            )
-
-            return primal_out, tangent_out
-
-        return f(value, self.alpha, self.low, self.high)
-
-    def icdf(self, q):
-        r"""Inverse cumulated probability distribution:
-        Z inequal minus one:
-
-        .. math::
-            a \left(\frac{b}{a}\right)^{q}
-
-        Z equal minus one:
-
-        .. math::
-            \left(a^{1 + \alpha} + q (b^{1 + \alpha} - a^{1 + \alpha})\right)^{\frac{1}{1 + \alpha}}
-
-        Derivations are calculated by Wolfram Alpha via the Jacobian matrix accordingly.
-        """
-
-        @jax.custom_jvp
-        def f(q, alpha, low, high):
-            neq_neg1_mask = jnp.not_equal(alpha, -1.0)
-            neq_neg1_alpha = jnp.where(neq_neg1_mask, alpha, 0.0)
-
-            def icdf_alpha_neq_neg1():
-                one_more_alpha = 1.0 + neq_neg1_alpha
-                low_pow_one_more_alpha = jnp.power(low, one_more_alpha)
-                high_pow_one_more_alpha = jnp.power(high, one_more_alpha)
-                return jnp.power(
-                    low_pow_one_more_alpha
-                    + q * (high_pow_one_more_alpha - low_pow_one_more_alpha),
-                    jnp.reciprocal(one_more_alpha),
-                )
-
-            def icdf_alpha_eq_neg1():
-                return jnp.power(high / low, q) * low
-
-            icdf_val = jnp.where(
-                neq_neg1_mask,
-                icdf_alpha_neq_neg1(),
-                icdf_alpha_eq_neg1(),
-            )
-            return icdf_val
-
-        @f.defjvp
-        def f_jvp(primals, tangents):
-            x, alpha, low, high = primals
-            x_t, alpha_t, low_t, high_t = tangents
-
-            log_low = jnp.log(low)
-            log_high = jnp.log(high)
-            high_over_low = jnp.divide(high, low)
-
-            delta_eq_neg1 = 10e-4
-            neq_neg1_mask = jnp.not_equal(alpha, -1.0)
-            neq_neg1_alpha = jnp.where(neq_neg1_mask, alpha, 0.0)
-
-            primal_out = f(*primals)
-
-            # Tangents for alpha not equal -1
-            def x_neq_neg1(alpha):
-                one_more_alpha = 1.0 + alpha
-                low_pow_one_more_alpha = jnp.power(low, one_more_alpha)
-                high_pow_one_more_alpha = jnp.power(high, one_more_alpha)
-                change = high_pow_one_more_alpha - low_pow_one_more_alpha
-                return (
-                    change
-                    * jnp.power(
-                        low_pow_one_more_alpha + x * change,
-                        jnp.reciprocal(one_more_alpha) - 1,
-                    )
-                ) / one_more_alpha
-
-            def alpha_neq_neg1(alpha):
-                one_more_alpha = 1.0 + alpha
-                low_pow_one_more_alpha = jnp.power(low, one_more_alpha)
-                high_pow_one_more_alpha = jnp.power(high, one_more_alpha)
-                factor0 = low_pow_one_more_alpha + x * (
-                    high_pow_one_more_alpha - low_pow_one_more_alpha
-                )
-                term1 = jnp.power(factor0, jnp.reciprocal(one_more_alpha))
-                term2 = (
-                    low_pow_one_more_alpha * log_low
-                    + x
-                    * (
-                        high_pow_one_more_alpha * log_high
-                        - low_pow_one_more_alpha * log_low
-                    )
-                ) / (one_more_alpha * factor0)
-                term3 = jnp.log(factor0) / jnp.square(one_more_alpha)
-                return term1 * (term2 - term3)
-
-            def low_neq_neg1(alpha):
-                one_more_alpha = 1.0 + alpha
-                low_pow_one_more_alpha = jnp.power(low, one_more_alpha)
-                high_pow_one_more_alpha = jnp.power(high, one_more_alpha)
-                return (
-                    (1.0 - x)
-                    * jnp.power(low, alpha)
-                    * jnp.power(
-                        low_pow_one_more_alpha
-                        + x * (high_pow_one_more_alpha - low_pow_one_more_alpha),
-                        jnp.reciprocal(one_more_alpha) - 1,
-                    )
-                )
-
-            def high_neq_neg1(alpha):
-                one_more_alpha = 1.0 + alpha
-                low_pow_one_more_alpha = jnp.power(low, one_more_alpha)
-                high_pow_one_more_alpha = jnp.power(high, one_more_alpha)
-                return (
-                    x
-                    * jnp.power(high, alpha)
-                    * jnp.power(
-                        low_pow_one_more_alpha
-                        + x * (high_pow_one_more_alpha - low_pow_one_more_alpha),
-                        jnp.reciprocal(one_more_alpha) - 1,
-                    )
-                )
-
-            # Tangents for alpha equals -1
-            def dx_eq_neg1():
-                return low * jnp.power(high_over_low, x) * (log_high - log_low)
-
-            def low_eq_neg1():
-                return (
-                    jnp.power(high_over_low, x)
-                    - (high * x * jnp.power(high_over_low, x - 1)) / low
-                )
-
-            def high_eq_neg1():
-                return x * jnp.power(high_over_low, x - 1)
-
-            # Inlcuding approximation for alpha = -1 \
-            tangent_out = (
-                jnp.where(neq_neg1_mask, x_neq_neg1(neq_neg1_alpha), dx_eq_neg1()) * x_t
-                + jnp.where(
-                    neq_neg1_mask,
-                    alpha_neq_neg1(neq_neg1_alpha),
-                    (
-                        alpha_neq_neg1(alpha - delta_eq_neg1)
-                        + alpha_neq_neg1(alpha + delta_eq_neg1)
-                    )
-                    * 0.5,
-                )
-                * alpha_t
-                + jnp.where(neq_neg1_mask, low_neq_neg1(neq_neg1_alpha), low_eq_neg1())
-                * low_t
-                + jnp.where(
-                    neq_neg1_mask, high_neq_neg1(neq_neg1_alpha), high_eq_neg1()
-                )
-                * high_t
-            )
-
-            return primal_out, tangent_out
-
-        return f(q, self.alpha, self.low, self.high)
-
-    def sample(self, key, sample_shape=()):
-        assert is_prng_key(key)
-        u = random.uniform(key, sample_shape + self.batch_shape)
-        samples = self.icdf(u)
-        return samples
-
-
-class LowerTruncatedPowerLaw(Distribution):
-    r"""Lower truncated power law distribution with :math:`\alpha` index.
-    We can define the power law distribution as,
-
-    .. math::
-        f(x; \alpha, a) = (-\alpha-1)a^{-\alpha - 1}x^{-\alpha},
-        \qquad x \geq a, \qquad \alpha < -1,
-
-    where, :math:`a` is the lower bound. The cdf of the distribution is given by,
-
-    .. math::
-        F(x; \alpha, a) = 1 - \left(\frac{x}{a}\right)^{1+\alpha}.
-
-    The k-th moment of the distribution is given by,
-
-    .. math::
-        E[X^k] = \begin{cases}
-            \frac{-\alpha-1}{-\alpha-1-k}a^k & \text{if } k < -\alpha-1, \\
-            \infty & \text{otherwise}.
-        \end{cases}
-
-    :param alpha: index of the power law distribution
-    :param low: lower bound of the distribution
-    """
-
-    arg_constraints = {
-        "alpha": constraints.less_than(-1.0),
-        "low": constraints.greater_than(0.0),
-    }
-    reparametrized_params = ["alpha", "low"]
-    pytree_aux_fields = ("_support",)
-
-    def __init__(self, alpha, low, *, validate_args=None):
-        self.alpha, self.low = promote_shapes(alpha, low)
-        batch_shape = lax.broadcast_shapes(jnp.shape(alpha), jnp.shape(low))
-        self._support = constraints.greater_than(low)
-        super(LowerTruncatedPowerLaw, self).__init__(
-            batch_shape=batch_shape, validate_args=validate_args
-        )
-
-    @constraints.dependent_property(is_discrete=False, event_dim=0)
-    def support(self):
-        return self._support
-
-    @validate_sample
-    def log_prob(self, value):
-        one_more_alpha = 1.0 + self.alpha
-        return (
-            self.alpha * jnp.log(value)
-            + jnp.log(-one_more_alpha)
-            - one_more_alpha * jnp.log(self.low)
-        )
-
-    def cdf(self, value):
-        cdf_val = jnp.where(
-            jnp.less_equal(value, self.low),
-            jnp.zeros_like(value),
-            1.0 - jnp.power(value / self.low, 1.0 + self.alpha),
-        )
-        return cdf_val
-
-    def icdf(self, q):
-        nan_mask = jnp.logical_or(jnp.isnan(q), jnp.less(q, 0.0))
-        nan_mask = jnp.logical_or(nan_mask, jnp.greater(q, 1.0))
-        return jnp.where(
-            nan_mask,
-            jnp.nan,
-            self.low * jnp.power(1.0 - q, jnp.reciprocal(1.0 + self.alpha)),
-        )
-
-    def sample(self, key, sample_shape=()):
-        assert is_prng_key(key)
-        u = random.uniform(key, sample_shape + self.batch_shape)
-        samples = self.icdf(u)
-        return samples
